DP0001 Finished, database H2 setup, applied documentation architecture layout, added BusinessApplicationService, changed Service return types from entities to DTOs
