+++ src/main/java/com/example/demo/BusinessComponent/API/DTOs/CreateBusinessDTO.java
@@ -1,4 +1,4 @@
-package com.example.demo.BusinessSystem.DTOs;
+package com.example.demo.BusinessComponent.API.DTOs;
 
 import java.util.UUID;
 
@@ -11,9 +11,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class BusinessDTO {
-  @NotNull
-  private UUID id;
+public class CreateBusinessDTO {
   @NotNull
   private String name;
   @NotNull
