+++ src/main/java/com/example/demo/BusinessComponent/Domain/Entities/Business.java
@@ -1,13 +1,9 @@
-package com.example.demo.BusinessSystem.Entities;
+package com.example.demo.BusinessComponent.Domain.Entities;
 
 import java.util.UUID;
 
-import com.example.demo.BusinessSystem.Enums.RoleType;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -28,22 +24,21 @@ import lombok.Setter;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "\"user\"")
-public class User {
+@Table(name = "business")
+public class Business {
   @Id
   @GeneratedValue(strategy = GenerationType.UUID)
   private UUID id;
 
-  @Column(name = "full_name", nullable = false)
-  private String fullName;
+  @Column(name = "name", nullable = false)
+  private String name;
 
   @OneToOne
-  @JoinColumn(name = "business_id", referencedColumnName = "id", nullable = true)
-  private Business business;
+  @JoinColumn(name = "owner_id", referencedColumnName = "id", nullable = false)
+  private User owner;
 
-  @Enumerated(EnumType.STRING)
-  @Column(name = "role", nullable = false)
-  private RoleType role;
+  @Column(name = "address", nullable = false)
+  private String address;
 
   @Pattern(regexp = "\\+\\d{3,30}")
   @Column(name = "phone_number", nullable = false)
