+++ src/main/java/com/example/demo/BusinessComponent/Domain/Entities/User.java
@@ -1,9 +1,13 @@
-package com.example.demo.BusinessSystem.Entities;
+package com.example.demo.BusinessComponent.Domain.Entities;
 
 import java.util.UUID;
 
+import com.example.demo.BusinessComponent.Domain.Entities.Enums.RoleType;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -24,21 +28,22 @@ import lombok.Setter;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "business")
-public class Business {
+@Table(name = "\"user\"")
+public class User {
   @Id
   @GeneratedValue(strategy = GenerationType.UUID)
   private UUID id;
 
-  @Column(name = "name", nullable = false)
-  private String name;
+  @Column(name = "full_name", nullable = false)
+  private String fullName;
 
   @OneToOne
-  @JoinColumn(name = "owner_id", referencedColumnName = "id", nullable = false)
-  private User owner;
+  @JoinColumn(name = "business_id", referencedColumnName = "id", nullable = true)
+  private Business business;
 
-  @Column(name = "address", nullable = false)
-  private String address;
+  @Enumerated(EnumType.STRING)
+  @Column(name = "role", nullable = false)
+  private RoleType role;
 
   @Pattern(regexp = "\\+\\d{3,30}")
   @Column(name = "phone_number", nullable = false)
