+++ src/main/java/com/example/demo/BusinessComponent/Repositories/IBusinessRepository.java
@@ -1,10 +1,10 @@
-package com.example.demo.BusinessSystem.Repositories;
+package com.example.demo.BusinessComponent.Repositories;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.BusinessSystem.Entities.Business;
+import com.example.demo.BusinessComponent.Domain.Entities.Business;
 
 public interface IBusinessRepository extends JpaRepository<Business, UUID> {
 }
