+++ src/main/java/com/example/demo/BusinessComponent/Repositories/IUserRepository.java
@@ -1,10 +1,10 @@
-package com.example.demo.BusinessSystem.Repositories;
+package com.example.demo.BusinessComponent.Repositories;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.BusinessSystem.Entities.User;
+import com.example.demo.BusinessComponent.Domain.Entities.User;
 
 public interface IUserRepository extends JpaRepository<User, UUID> {
 }
