--- src/main/java/com/example/demo/BusinessSystem/Mappers/Interfaces/IBusinessMapper.java
@@ -1,10 +1,0 @@
-package com.example.demo.BusinessSystem.Mappers.Interfaces;
-
-import com.example.demo.BusinessSystem.DTOs.BusinessDTO;
-import com.example.demo.BusinessSystem.DTOs.CreateBusinessDTO;
-import com.example.demo.BusinessSystem.Entities.Business;
-
-public interface IBusinessMapper {
-  public BusinessDTO businessToBusinessDTO(Business business);
-  public Business createBusinessDTOToBusiness(CreateBusinessDTO createBusinessDTO);
-}
