--- src/main/java/com/example/demo/BusinessSystem/Services/Interfaces/IBusinessService.java
@@ -1,8 +1,0 @@
-package com.example.demo.BusinessSystem.Services.Interfaces;
-
-import com.example.demo.BusinessSystem.DTOs.CreateBusinessDTO;
-import com.example.demo.BusinessSystem.Entities.Business;
-
-public interface IBusinessService {
-  public Business createBusiness(CreateBusinessDTO dto);
-}
